Replace all game art

+++ src/__main__.py
@@ -5,53 +5,67 @@ import shutil
 import tkinter as tk
 import urllib.request
 from pathlib import Path
-from tkinter import Frame, ttk
+from tkinter import ttk
 
 import booru
 
-# TODO: There can be other paths, maybe multiple?
-match platform.system():
-    case "Linux":
-        STEAM_PATH = Path("~/.local/share/Steam").expanduser()
-    case "Windows":
-        STEAM_PATH = Path("C:/Program Files (x86)/Steam")
 
-USER_PATH_LIST = list((STEAM_PATH / "userdata").iterdir())
+def get_steam_path() -> Path:
+    # TODO: There can be other paths, maybe multiple?
+    match platform.system():
+        case "Linux":
+            return Path("~/.local/share/Steam").expanduser()
+        case "Windows":
+            return Path("C:/Program Files (x86)/Steam")
 
 
-async def pornify(user: str) -> None:
-    # TODO: There can be multiple users
-    grid_path = STEAM_PATH / "userdata" / user / "config" / "grid"
+def paths_to_ids(dir: Path) -> list[str]:
+    return [path.name for path in dir.iterdir() if path.is_dir()]
+
+
+def get_game_ids() -> list[str]:
+    return paths_to_ids(get_steam_path() / "appcache" / "librarycache")
+
+
+def get_user_ids() -> list[Path]:
+    return paths_to_ids(get_steam_path() / "userdata")
+
+
+def get_grid_path(user_id: str) -> None:
+    return get_steam_path() / "userdata" / user_id / "config" / "grid"
+
+
+async def pornify(user_id: str) -> None:
+    grid_path = get_grid_path(user_id)
     grid_path.mkdir(parents=True, exist_ok=True)
 
     dan = booru.Danbooru()
     res = await dan.search_image(query="order:rank")
     image = random.choice(booru.resolve(res))
-    urllib.request.urlretrieve(image, grid_path / "4000_hero.png")
+    urllib.request.urlretrieve(image, grid_path / "image.png")
+
+    for game_id in get_game_ids():
+        for art_suffix in ["", "p", "_hero"]:
+            shutil.copyfile(grid_path / "image.png", grid_path / f"{game_id}{art_suffix}.png")
 
     print("Done Pornify")
 
-def resteam(user: str) -> None:
-    # TODO: There can be multiple users
-    grid_path = STEAM_PATH / "userdata" / user / "config" / "grid"
-    grid_path.mkdir(parents=True, exist_ok=True)
 
-    shutil.rmtree(grid_path)
+def resteam(user_id: str) -> None:
+    grid_path = get_grid_path(user_id)
+    if grid_path.exists():
+        shutil.rmtree(grid_path)
     print("Done Resteam")
+
 
 if __name__ == "__main__":
     root = tk.Tk()
     root.geometry("320x240")
 
-    # TODO: Get actual username corresponding to user id and display it here for ease of use
-    usernames = []
-    for user_path in USER_PATH_LIST:
-        usernames.append(user_path.name)
-
-    run_frame = Frame(root)
+    run_frame = tk.Frame(root)
     run_frame.pack(expand=1)
 
-    user_select_frame = Frame(run_frame)
+    user_select_frame = tk.Frame(run_frame)
     user_select_frame.pack()
 
     # For making sure that the "Select Steam Account" folder isn't accidentally created
@@ -60,17 +74,18 @@ if __name__ == "__main__":
             pornify_button.config(state="normal")
             resteam_button.config(state="normal")
 
+    # TODO: Get actual username corresponding to user id and display it here for ease of use
     user_var = tk.StringVar(root)
     user_var.trace("w", enable_buttons)
-    user_selection_dropdown = ttk.Combobox(user_select_frame, state="readonly", textvariable=user_var, values=usernames)
+    user_selection_dropdown = ttk.Combobox(user_select_frame, state="readonly", textvariable=user_var, values=get_user_ids())
     user_selection_dropdown.set("Select Steam Account")
     user_selection_dropdown.pack(pady=5)
 
-    start_stop_frame = Frame(run_frame)
+    start_stop_frame = tk.Frame(run_frame)
     start_stop_frame.pack()
-    pornify_button = ttk.Button(start_stop_frame, text="Pornify", command=lambda: asyncio.run(pornify(user_selection_dropdown.get())))
+    pornify_button = ttk.Button(start_stop_frame, text="Pornify", command=lambda: asyncio.run(pornify(user_var.get())))
     pornify_button.grid(row=0, column=0, ipady=5)
-    resteam_button = ttk.Button(start_stop_frame, text="Resteam", command=lambda: resteam(user_selection_dropdown.get()))
+    resteam_button = ttk.Button(start_stop_frame, text="Resteam", command=lambda: resteam(user_var.get()))
     resteam_button.grid(row=0, column=1, ipady=5)
     # Initially disable buttons to wait for user to be properly selected
     pornify_button.config(state="disabled")
